Move classes to packages, fix logger exception (unexpected '{') - wrong formatting of string
