+++ main/core/database/benchmark_administration.py
@@ -4,7 +4,7 @@ from os.path import isfile
 from main.core.database import groups
 from main.core.database.db import Database, DatabaseException
 from main.core.hashing.gbd_hash import gbd_hash
-from main.core.util import eprint
+from main.util import eprint
 
 
 def add_tag(database, cat, tag, hash, force=False):
