+++ main/core/database/db.py
@@ -2,7 +2,7 @@ import sqlite3
 from os.path import isfile
 
 from main.core.hashing.gbd_hash import HASH_VERSION
-from main.core.util import eprint
+from main.util import eprint
 
 VERSION = 0
